refactor(driver): extract aarch64 window search inputs into a pure tested seam
Lift the four pure AArch64 per-window "search inputs" rules out of the
6136-line `main.rs` binary driver into a new library module,
`src/aarch64_search_inputs.rs`. This mirrors the x86 extraction in #752
(`src/x86_search_inputs.rs`) and restores AArch64/x86 parity — the x86
module already referenced "the AArch64 builder" as a sibling seam that
did not yet exist.

Extracted, moved verbatim (name/visibility/doc only):

- `registers_from_target` (was `aarch64_search_registers`)
- `live_out_for_optimization_prefix`
- `validate_basic_block`

Plus `default_immediates()`, which names and tests the AArch64 immediate
pool that was previously an untested literal inline in `run_optimization`.

The driver becomes a thin adapter that calls the seam. Each rule is now a
fixture-free unit test instead of requiring a whole search to exercise;
the moved unit tests come with the functions, and two integration tests
in `main.rs` now call the namespaced seam. New coverage: the immediate
pool and the X0..X7 scalar-base ordering, neither tested before.

This is one step in an ongoing sequence of driver extractions, not a
wholesale fix of `main.rs`. The `build_*_search_config` builders stay in
`main.rs` because they depend on the binary-private `OptimizationOptions`,
matching the boundary #752 drew for the x86 builders.
