+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod aarch64_search_inputs;
 pub mod assembler;
 pub mod bench_support;
 pub mod candidate_windows;
