+++ src/main.rs
@@ -22,11 +22,11 @@ use s11::search::config::{
 };
 use s11::search::parallel::{ParallelConfig, run_parallel_search};
 use s11::search::{EnumerativeSearch, SearchAlgorithm, StochasticSearch, SymbolicSearch};
-use s11::semantics::LiveOut;
 use s11::semantics::cost::CostMetric;
 #[allow(unused_imports)]
 use s11::{
-    assembler, elf_patcher, ir, isa, parser, search, semantics, validation, x86_search_inputs,
+    aarch64_search_inputs, assembler, elf_patcher, ir, isa, parser, search, semantics, validation,
+    x86_search_inputs,
 };
 
 // --- Command Line Arguments ---
@@ -745,7 +745,7 @@ impl ElfOptimizationBackend for AArch64OptimizationBackend {
     }
 
     fn validate_window_ir(&self, ir: &[Self::Instruction]) -> Result<(), String> {
-        validate_basic_block(ir)
+        aarch64_search_inputs::validate_basic_block(ir)
     }
 
     fn optimization_context(
@@ -1612,45 +1612,6 @@ fn optimize_elf_binary_with_backend<B: ElfOptimizationBackend>(
 /// mirroring the `flags_live = if terminator.is_some() { true }` blanket. When
 /// a terminator is present we ignore `downstream_live` and keep every
 /// window-written register live.
-fn live_out_for_optimization_prefix(
-    prefix: &[Instruction],
-    terminator: Option<&Instruction>,
-    downstream_flags_live: bool,
-    downstream_live: Option<&semantics::live_out::RegisterSet<Register>>,
-) -> LiveOut {
-    // A terminator vetoes register narrowing (its other successor is unscanned).
-    let narrowing = if terminator.is_some() {
-        None
-    } else {
-        downstream_live
-    };
-
-    let mut live_registers: Vec<Register> = match narrowing {
-        // Narrow to (written ∩ proven-live). The downstream set is already a
-        // subset of the window-written registers (it is computed from exactly
-        // that candidate set), so iterating it is sufficient.
-        Some(live) => live.iter().copied().collect(),
-        // No downstream analysis (or vetoed by a terminator): keep every
-        // written register live.
-        None => prefix
-            .iter()
-            .flat_map(|instr| instr.destinations())
-            .collect(),
-    };
-
-    if let Some(terminator) = terminator {
-        live_registers.extend(terminator.source_registers());
-    }
-
-    let flags_live = if terminator.is_some() {
-        true
-    } else {
-        downstream_flags_live
-    };
-
-    LiveOut::from_registers(live_registers).with_flags(flags_live)
-}
-
 /// Shared base `SearchConfig` for the AArch64 stochastic/enumerative/hybrid/
 /// symbolic/LLM builders. Sets the fields every AArch64 algorithm configures
 /// identically — cost metric, overall and SMT solver timeouts, verbosity, and
@@ -1799,40 +1760,6 @@ fn build_x86_symbolic_search_config(
         .with_x86_same_count_code_size_allowed(same_count_code_size_allowed)
 }
 
-/// Build the AArch64 search register pool, preserving the historical X0..X7
-/// scalar policy while adding every vector register referenced by the target.
-/// A target-local vector pool avoids the 32^3 candidate explosion of enabling
-/// the entire SIMD register file for small enumerative searches.
-fn aarch64_search_registers(target: &[Instruction]) -> Vec<Register> {
-    let mut registers = vec![
-        Register::X0,
-        Register::X1,
-        Register::X2,
-        Register::X3,
-        Register::X4,
-        Register::X5,
-        Register::X6,
-        Register::X7,
-    ];
-
-    for register in target
-        .iter()
-        .flat_map(|instruction| instruction.source_registers().into_iter())
-        .chain(
-            target
-                .iter()
-                .flat_map(|instruction| instruction.destinations().into_iter()),
-        )
-        .filter(|register| register.vector().is_some())
-    {
-        if !registers.contains(&register) {
-            registers.push(register);
-        }
-    }
-    registers.sort_by_key(|register| register.sort_key());
-    registers
-}
-
 /// Run optimization using the selected algorithm.
 ///
 /// Issue #69: if `target` ends in a terminator (branch / control-flow
@@ -1859,16 +1786,14 @@ fn run_optimization(
 
     // Keep the historical scalar pool and add the target's vector registers
     // so every search backend can generate NEON candidates for NEON windows.
-    let available_registers = aarch64_search_registers(prefix);
-    let available_immediates = vec![
-        0, 1, 2, 3, 4, 5, 7, 8, 10, 15, 16, 31, 32, 63, 64, 100, 255, 256, 1000, 4095,
-    ];
+    let available_registers = aarch64_search_inputs::registers_from_target(prefix);
+    let available_immediates = aarch64_search_inputs::default_immediates();
 
     // Create live-out contract over the prefix (assume all modified registers
     // are live-out), plus any registers the fixed terminator reads after the
     // optimized prefix runs. NZCV liveness comes from the fixed terminator or
     // the known downstream fall-through context.
-    let live_out = live_out_for_optimization_prefix(
+    let live_out = aarch64_search_inputs::live_out_for_optimization_prefix(
         prefix,
         terminator,
         downstream_flags_live,
@@ -2139,25 +2064,6 @@ fn optimization_context_for_backend(
     }
 
     OptimizationContext::default()
-}
-
-/// Validate that an IR sequence forms a single basic block: at most one
-/// terminator (branch / control-flow instruction), and only at the final
-/// position. Issue #69 scope — internal branches mid-block are rejected.
-///
-/// Accepted shapes: `[]`, `[i1, ..., ik]` (no branch), `[t]` (terminator
-/// only), `[i1, ..., ik, t]` (prefix + terminator).
-fn validate_basic_block(ir: &[Instruction]) -> Result<(), String> {
-    let last_idx = ir.len().saturating_sub(1);
-    for (i, instr) in ir.iter().enumerate() {
-        if i < last_idx && instr.is_terminator() {
-            return Err(format!(
-                "Region contains a branch at position {} ({}); only single basic blocks ending in a terminator are supported (issue #69 scope)",
-                i, instr
-            ));
-        }
-    }
-    Ok(())
 }
 
 // ============================================================================
@@ -4520,107 +4426,6 @@ mod cli_helper_tests {
     }
 
     #[test]
-    fn live_out_for_optimization_prefix_narrows_to_downstream_live_regs() {
-        // Prefix writes x0 and x1.
-        let prefix = [
-            Instruction::MovImm {
-                rd: Register::X0,
-                imm: 0,
-            },
-            Instruction::MovImm {
-                rd: Register::X1,
-                imm: 0,
-            },
-        ];
-
-        // Default (no downstream analysis): both written registers stay live.
-        let default = live_out_for_optimization_prefix(&prefix, None, false, None);
-        assert!(default.contains(Register::X0));
-        assert!(default.contains(Register::X1));
-
-        // Downstream scan proved only x1 live (x0 dead): drop x0, pin x1.
-        let downstream_live = semantics::live_out::RegisterSet::from_registers(vec![Register::X1]);
-        let narrowed =
-            live_out_for_optimization_prefix(&prefix, None, false, Some(&downstream_live));
-        assert!(
-            !narrowed.contains(Register::X0),
-            "a provably-dead window register must be dropped from live-out"
-        );
-        assert!(
-            narrowed.contains(Register::X1),
-            "a downstream-live window register must stay pinned"
-        );
-    }
-
-    /// Soundness regression: a window whose held-fixed terminator is a
-    /// CONDITIONAL branch must NOT narrow window-written registers, even if the
-    /// linear fall-through suffix proved one dead. The downstream-regs scan only
-    /// follows the fall-through successor; the branch-TAKEN successor is never
-    /// inspected and may read the register's window value.
-    ///
-    /// Counterexample being guarded against:
-    ///   window:       mov x0, #7 ; b.eq TARGET
-    ///   fall-through: mov x0, #0 ; ret           (kills x0 -> scan says Dead)
-    ///   elsewhere:    TARGET: add x9, x0, #1     (READS x0 on the taken path)
-    /// If x0 were narrowed to dead, `mov x0, #7` could be deleted and the
-    /// b.eq-taken path would read a stale x0. `BCond::source_registers()` is
-    /// empty, so the terminator does not re-pin x0 either — the only correct
-    /// fix is to not narrow at all when a terminator is present.
-    #[test]
-    fn live_out_for_optimization_prefix_does_not_narrow_with_conditional_terminator() {
-        let prefix = [Instruction::MovImm {
-            rd: Register::X0,
-            imm: 7,
-        }];
-        let b_eq = Instruction::BCond {
-            target: s11::ir::LabelId(0x2000),
-            cond: s11::ir::Condition::EQ,
-        };
-
-        // The fall-through scan "proved" x0 dead (empty proven-live set).
-        let downstream_live_fall_through = semantics::live_out::RegisterSet::<Register>::empty();
-
-        let live_out = live_out_for_optimization_prefix(
-            &prefix,
-            Some(&b_eq),
-            false,
-            Some(&downstream_live_fall_through),
-        );
-
-        assert!(
-            live_out.contains(Register::X0),
-            "x0 must stay live: a conditional terminator has a branch-taken successor \
-             the fall-through scan never inspected, so register narrowing must not apply"
-        );
-    }
-
-    /// Same soundness gate for unconditional terminators: the instruction at
-    /// `end_addr` is not the real/only successor, so narrowing must not apply.
-    #[test]
-    fn live_out_for_optimization_prefix_does_not_narrow_with_unconditional_terminator() {
-        let prefix = [Instruction::MovImm {
-            rd: Register::X0,
-            imm: 7,
-        }];
-        let cases = [
-            Instruction::B {
-                target: s11::ir::LabelId(0x2000),
-            },
-            Instruction::Ret { rn: Register::X30 },
-        ];
-        let dead = semantics::live_out::RegisterSet::<Register>::empty();
-        for terminator in cases {
-            let live_out =
-                live_out_for_optimization_prefix(&prefix, Some(&terminator), false, Some(&dead));
-            assert!(
-                live_out.contains(Register::X0),
-                "x0 must stay live with a {:?} terminator: narrowing must not apply",
-                terminator
-            );
-        }
-    }
-
-    #[test]
     fn x86_symbolic_code_size_preserves_downstream_flags_live() {
         let mut opts = options_for(Algorithm::Symbolic);
         opts.timeout = Some(Duration::from_secs(5));
@@ -5213,36 +5018,6 @@ mod cli_helper_tests {
         assert_eq!(config.cores, SearchConfig::default().cores);
     }
 
-    #[test]
-    fn aarch64_search_registers_include_vectors_used_by_target() {
-        let target = [
-            Instruction::VectorAdd {
-                vd: ir::VectorRegister::V0,
-                vn: ir::VectorRegister::V1,
-                vm: ir::VectorRegister::V2,
-                arrangement: ir::VectorArrangement::TwoD,
-            },
-            Instruction::MovFromVectorLane {
-                rd: Register::X0,
-                vn: ir::VectorRegister::V0,
-                lane: 0,
-            },
-        ];
-
-        let registers = aarch64_search_registers(&target);
-
-        assert!(registers.contains(&Register::Vector(ir::VectorRegister::V0)));
-        assert!(registers.contains(&Register::Vector(ir::VectorRegister::V1)));
-        assert!(registers.contains(&Register::Vector(ir::VectorRegister::V2)));
-        assert_eq!(
-            registers
-                .iter()
-                .filter(|reg| reg.vector().is_some())
-                .count(),
-            3
-        );
-    }
-
     /// Regression for issue #243, generalised: every AArch64 algorithm builder
     /// must propagate the shared base fields (`--timeout`, `--solver-timeout`,
     /// cost metric, verbosity, register/immediate pools) identically. They all
@@ -5526,47 +5301,6 @@ mod cli_helper_tests {
         run_llm_opt(llm_asm.path(), "x0", 0, "test-model", 0, true).unwrap();
     }
 
-    // ===== Issue #69: validate_basic_block =====
-
-    #[test]
-    fn validate_basic_block_accepts_empty_sequence() {
-        assert!(validate_basic_block(&[]).is_ok());
-    }
-
-    #[test]
-    fn validate_basic_block_accepts_prefix_only_no_terminator() {
-        let seq = vec![
-            Instruction::MovImm {
-                rd: Register::X0,
-                imm: 1,
-            },
-            Instruction::Add {
-                rd: Register::X1,
-                rn: Register::X0,
-                rm: Operand::Immediate(2),
-            },
-        ];
-        assert!(validate_basic_block(&seq).is_ok());
-    }
-
-    #[test]
-    fn validate_basic_block_accepts_terminator_only() {
-        let seq = vec![Instruction::Ret { rn: Register::X30 }];
-        assert!(validate_basic_block(&seq).is_ok());
-    }
-
-    #[test]
-    fn validate_basic_block_accepts_prefix_plus_terminator() {
-        let seq = vec![
-            Instruction::MovImm {
-                rd: Register::X0,
-                imm: 1,
-            },
-            Instruction::Ret { rn: Register::X30 },
-        ];
-        assert!(validate_basic_block(&seq).is_ok());
-    }
-
     #[test]
     fn split_terminator_returns_full_slice_when_no_terminator() {
         let seq = vec![Instruction::MovImm {
@@ -5590,78 +5324,6 @@ mod cli_helper_tests {
         let (prefix, term) = split_terminator(&seq);
         assert_eq!(prefix.len(), 1);
         assert_eq!(term, Some(&Instruction::Ret { rn: Register::X30 }));
-    }
-
-    #[test]
-    fn live_out_for_optimization_prefix_includes_registers_read_by_terminator() {
-        let prefix = [Instruction::MovImm {
-            rd: Register::X1,
-            imm: 1,
-        }];
-        let cases = [
-            (
-                Instruction::Cbz {
-                    rn: Register::X0,
-                    target: s11::ir::LabelId(0x1000),
-                },
-                Register::X0,
-            ),
-            (
-                Instruction::Tbz {
-                    rt: Register::X2,
-                    bit: 5,
-                    target: s11::ir::LabelId(0x1000),
-                },
-                Register::X2,
-            ),
-            (Instruction::Br { rn: Register::X16 }, Register::X16),
-            (Instruction::Ret { rn: Register::X30 }, Register::X30),
-        ];
-
-        for (terminator, source) in cases {
-            let live_out =
-                live_out_for_optimization_prefix(&prefix, Some(&terminator), false, None);
-            assert!(live_out.contains_register(Register::X1));
-            assert!(
-                live_out.contains_register(source),
-                "{:?} must keep {:?} live for the reattached terminator",
-                terminator,
-                source
-            );
-        }
-    }
-
-    #[test]
-    fn live_out_for_optimization_prefix_uses_downstream_flags_without_terminator() {
-        let prefix = [Instruction::MovImm {
-            rd: Register::X1,
-            imm: 1,
-        }];
-
-        let flags_dead = live_out_for_optimization_prefix(&prefix, None, false, None);
-        assert!(!flags_dead.flags_live());
-
-        let flags_live = live_out_for_optimization_prefix(&prefix, None, true, None);
-        assert!(flags_live.flags_live());
-    }
-
-    #[test]
-    fn live_out_for_optimization_prefix_keeps_flags_live_for_terminators() {
-        let prefix = [Instruction::MovImm {
-            rd: Register::X1,
-            imm: 1,
-        }];
-
-        let b_cond = Instruction::BCond {
-            target: s11::ir::LabelId(0x1000),
-            cond: s11::ir::Condition::EQ,
-        };
-        let live_out = live_out_for_optimization_prefix(&prefix, Some(&b_cond), false, None);
-        assert!(live_out.flags_live());
-
-        let ret = Instruction::Ret { rn: Register::X30 };
-        let live_out = live_out_for_optimization_prefix(&prefix, Some(&ret), false, None);
-        assert!(live_out.flags_live());
     }
 
     // (The standalone `find_shorter_equivalent_preserves_terminator_bit_identical`
@@ -5736,7 +5398,8 @@ mod cli_helper_tests {
         let (prefix, term) = split_terminator(&seq);
         assert_eq!(term, Some(&terminator), "split must recognize ret");
 
-        let live_out = live_out_for_optimization_prefix(prefix, term, true, None);
+        let live_out =
+            aarch64_search_inputs::live_out_for_optimization_prefix(prefix, term, true, None);
         let config = SearchConfig::default()
             .with_registers(vec![Register::X0, Register::X1])
             .with_immediates(vec![0, 1]);
@@ -5800,7 +5463,8 @@ mod cli_helper_tests {
         ];
 
         let (prefix, term) = split_terminator(&target);
-        let live_out = live_out_for_optimization_prefix(prefix, term, true, None);
+        let live_out =
+            aarch64_search_inputs::live_out_for_optimization_prefix(prefix, term, true, None);
         assert!(
             live_out.contains_register(Register::X0),
             "live_out_for_optimization_prefix must mark x0 live when the \
@@ -5818,30 +5482,6 @@ mod cli_helper_tests {
             "candidate that clobbers x0 must be rejected because the \
              reattached cbz reads x0; got {:?}",
             result,
-        );
-    }
-
-    #[test]
-    fn validate_basic_block_rejects_branch_mid_block() {
-        let seq = vec![
-            Instruction::MovImm {
-                rd: Register::X0,
-                imm: 1,
-            },
-            Instruction::B {
-                target: s11::ir::LabelId(0x1000),
-            },
-            Instruction::Add {
-                rd: Register::X1,
-                rn: Register::X0,
-                rm: Operand::Immediate(2),
-            },
-        ];
-        let err = validate_basic_block(&seq).expect_err("branch at position 1 must be rejected");
-        assert!(
-            err.contains("position 1") && err.contains("issue #69"),
-            "unexpected error: {}",
-            err
         );
     }
